requirements.txt
discord.py==2.4.0
pickleDB==0.9.2
pymongo==4.10.1
python-dotenv==1.0.1
Requests==2.32.3
thispersondoesnotexist==1.0.2
